+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/ExtensionScriptsUI.java
@@ -45,6 +45,7 @@ import org.parosproxy.paros.extension.ExtensionPopupMenuItem;
 import org.parosproxy.paros.extension.SessionChangedListener;
 import org.parosproxy.paros.model.Session;
 import org.parosproxy.paros.network.HttpMessage;
+import org.parosproxy.paros.view.OptionsDialog;
 import org.parosproxy.paros.view.View;
 import org.zaproxy.zap.ZAP;
 import org.zaproxy.zap.extension.api.API;
@@ -105,6 +106,9 @@ public class ExtensionScriptsUI extends ExtensionAdaptor implements ScriptEventL
     private ExtensionScript extScript = null;
     private ScriptsTreeCellRenderer renderer = null;
 
+    private ScriptConsoleOptions scriptConsoleOptions;
+    private ScriptConsoleOptionsPanel scriptConsoleOptionsPanel;
+
     private ScriptWrapper currentLockedScript = null;
     private boolean lockOutputToDisplayedScript = false;
 
@@ -140,6 +144,9 @@ public class ExtensionScriptsUI extends ExtensionAdaptor implements ScriptEventL
     @Override
     public void hook(ExtensionHook extensionHook) {
         super.hook(extensionHook);
+
+        extensionHook.addOptionsParamSet(getScriptConsoleOptions());
+
         this.getExtScript().addListener(this);
         extScriptType =
                 new ScriptType(
@@ -158,6 +165,12 @@ public class ExtensionScriptsUI extends ExtensionAdaptor implements ScriptEventL
         this.getExtScript().registerScriptEngineWrapper(nullEngineWrapper);
 
         if (hasView()) {
+            OptionsDialog optionsDialog = View.getSingleton().getOptionsDialog("");
+
+            String[] scriptNode = {Constant.messages.getString("options.script.title")};
+            scriptConsoleOptionsPanel = new ScriptConsoleOptionsPanel();
+            optionsDialog.addParamPanel(scriptNode, scriptConsoleOptionsPanel, true);
+
             extensionHook.getHookView().addSelectPanel(getScriptsPanel());
             extensionHook.addSessionListener(new ViewSessionChangedListener());
             extensionHook.getHookView().addWorkPanel(getConsolePanel());
@@ -252,6 +265,8 @@ public class ExtensionScriptsUI extends ExtensionAdaptor implements ScriptEventL
             if (scriptsPanel != null) {
                 scriptsPanel.unload();
             }
+            OptionsDialog optionsDialog = View.getSingleton().getOptionsDialog("");
+            optionsDialog.removeParamPanel(scriptConsoleOptionsPanel);
         }
 
         if (extScript != null) {
@@ -290,6 +305,13 @@ public class ExtensionScriptsUI extends ExtensionAdaptor implements ScriptEventL
             }
         }
         return extScript;
+    }
+
+    ScriptConsoleOptions getScriptConsoleOptions() {
+        if (scriptConsoleOptions == null) {
+            scriptConsoleOptions = new ScriptConsoleOptions();
+        }
+        return scriptConsoleOptions;
     }
 
     private ConsolePanel getConsolePanel() {