+++ addOns/scripts/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Allow selecting a default behaviour when a script in the console changes on disk (Issues 5463, 7582). The allowed
+  options are "Ask Each Time", "Keep Script", and "Replace Script".
+
 ### Changed
 - Update extender template scripts to also work with Graal.js engine.